Refactor base SDK modules
