+++ python/packages/sdk/pyproject.toml
@@ -32,6 +32,11 @@ documentation = "https://github.com/serverless/console/tree/main/python/packages
 homepage = "https://www.serverless.com/console"
 repository = "https://github.com/serverless/console"
 
+[tool.pytest.ini_options]
+testpaths = [
+    "tests",
+]
+
 [tool.ruff]
 ignore = ["F401"]
 #fix = true
