+++ python/packages/sdk/tests/test_trace_span.py
@@ -1,7 +1,7 @@
 import pytest
 import time
 from serverless_sdk.lib.timing import to_protobuf_epoch_timestamp
-from serverless_sdk.span.trace import TraceSpan
+from serverless_sdk.lib.trace import TraceSpan
 
 
 # root span that lives throughout the test session
@@ -198,25 +198,25 @@ def test_span_closure():
 def test_root_span_reuse():
     # given
     from importlib import reload
-    import serverless_sdk.span.trace
+    import serverless_sdk.lib.trace
 
-    reload(serverless_sdk.span.trace)
+    reload(serverless_sdk.lib.trace)
 
     span = TraceSpan("root")
     TraceSpan("child1").close()
     TraceSpan("child2").close()
     span.close()
     span.sub_spans.clear()
-    del serverless_sdk.span.trace.root_span.end_time
+    del serverless_sdk.lib.trace.root_span.end_time
 
     # when
     span.start_time = time.perf_counter_ns()
     TraceSpan("otherchild").close()
-    serverless_sdk.span.trace.root_span.close()
+    serverless_sdk.lib.trace.root_span.close()
 
     # then
-    assert [x.name for x in serverless_sdk.span.trace.root_span.spans] == [
+    assert [x.name for x in serverless_sdk.lib.trace.root_span.spans] == [
         "root",
         "otherchild",
     ]
-    serverless_sdk.span.trace.root_span.sub_spans.clear()
+    serverless_sdk.lib.trace.root_span.sub_spans.clear()
